+++ mycli/main.py
@@ -35,6 +35,7 @@ import click
 import clickdc
 from configobj import ConfigObj
 import keyring
+import prompt_toolkit
 from prompt_toolkit import print_formatted_text
 from prompt_toolkit.application.current import get_app
 from prompt_toolkit.auto_suggest import AutoSuggestFromHistory, ThreadedAutoSuggest
@@ -55,7 +56,8 @@ from prompt_toolkit.key_binding.key_processor import KeyPressEvent
 from prompt_toolkit.layout.processors import ConditionalProcessor, HighlightMatchingBracketProcessor
 from prompt_toolkit.lexers import PygmentsLexer
 from prompt_toolkit.output import ColorDepth
-from prompt_toolkit.shortcuts import CompleteStyle, PromptSession
+from prompt_toolkit.shortcuts import CompleteStyle, ProgressBar, PromptSession
+from prompt_toolkit.shortcuts.progress_bar import formatters as progress_bar_formatters
 import pymysql
 from pymysql.constants.CR import CR_SERVER_LOST
 from pymysql.constants.ER import ACCESS_DENIED_ERROR, HANDSHAKE_ERROR
@@ -2036,7 +2038,7 @@ class CliArgs:
     )
     ssl_verify_server_cert: bool = clickdc.option(
         is_flag=True,
-        help=('Verify server\'s "Common Name" in its cert against hostname used when connecting. This option is disabled by default.'),
+        help=("""Verify server's "Common Name" in its cert against hostname used when connecting. This option is disabled by default."""),
     )
     verbose: bool = clickdc.option(
         '-v',
@@ -2166,6 +2168,10 @@ class CliArgs:
         type=float,
         default=0.0,
         help='Pause in seconds between queries in batch mode.',
+    )
+    progress: bool = clickdc.option(
+        is_flag=True,
+        help='Show progress on the standard error with --batch.',
     )
     use_keyring: str | None = clickdc.option(
         type=click.Choice(['true', 'false', 'reset']),
@@ -2721,17 +2727,70 @@ def click_entrypoint(
             click.secho(str(e), err=True, fg="red")
             sys.exit(1)
 
-    if cli_args.batch or not sys.stdin.isatty():
-        if cli_args.batch:
-            if not sys.stdin.isatty() and cli_args.batch != '-':
-                click.secho('Ignoring STDIN since --batch was also given.', err=True, fg='red')
-            try:
-                batch_h = click.open_file(cli_args.batch)
-            except (OSError, FileNotFoundError):
-                click.secho(f'Failed to open --batch file: {cli_args.batch}', err=True, fg='red')
-                sys.exit(1)
-        else:
-            batch_h = click.get_text_stream('stdin')
+    if cli_args.batch and cli_args.batch != '-' and cli_args.progress and sys.stderr.isatty():
+        # The actual number of SQL statements can be greater, if there is more than
+        # one statement per line, but this is how the  progress bar will count.
+        goal_statements = 0
+        if not sys.stdin.isatty() and cli_args.batch != '-':
+            click.secho('Ignoring STDIN since --batch was also given.', err=True, fg='yellow')
+        if os.path.exists(cli_args.batch) and not os.path.isfile(cli_args.batch):
+            click.secho('--progress is only compatible with a plain file.', err=True, fg='red')
+            sys.exit(1)
+        try:
+            batch_count_h = click.open_file(cli_args.batch)
+            for _statement, _counter in statements_from_filehandle(batch_count_h):
+                goal_statements += 1
+            batch_count_h.close()
+            batch_h = click.open_file(cli_args.batch)
+        except (OSError, FileNotFoundError):
+            click.secho(f'Failed to open --batch file: {cli_args.batch}', err=True, fg='red')
+            sys.exit(1)
+        except ValueError as e:
+            click.secho(f'Error reading --batch file: {cli_args.batch}: {e}', err=True, fg='red')
+            sys.exit(1)
+        try:
+            if goal_statements:
+                pb_style = prompt_toolkit.styles.Style.from_dict({'bar-a': 'reverse'})
+                custom_formatters = [
+                    progress_bar_formatters.Bar(start='[', end=']', sym_a=' ', sym_b=' ', sym_c=' '),
+                    progress_bar_formatters.Text(' '),
+                    progress_bar_formatters.Progress(),
+                    progress_bar_formatters.Text(' '),
+                    progress_bar_formatters.Text('eta ', style='class:time-left'),
+                    progress_bar_formatters.TimeLeft(),
+                    progress_bar_formatters.Text(' ', style='class:time-left'),
+                ]
+                err_output = prompt_toolkit.output.create_output(stdout=sys.stderr, always_prefer_tty=True)
+                with ProgressBar(style=pb_style, formatters=custom_formatters, output=err_output) as pb:
+                    for pb_counter in pb(range(goal_statements)):
+                        statement, _untrusted_counter = next(statements_from_filehandle(batch_h))
+                        dispatch_batch_statements(statement, pb_counter)
+        except (ValueError, StopIteration) as e:
+            click.secho(str(e), err=True, fg='red')
+            sys.exit(1)
+        finally:
+            batch_h.close()
+        sys.exit(0)
+
+    if cli_args.batch:
+        if not sys.stdin.isatty() and cli_args.batch != '-':
+            click.secho('Ignoring STDIN since --batch was also given.', err=True, fg='red')
+        try:
+            batch_h = click.open_file(cli_args.batch)
+        except (OSError, FileNotFoundError):
+            click.secho(f'Failed to open --batch file: {cli_args.batch}', err=True, fg='red')
+            sys.exit(1)
+        try:
+            for statement, counter in statements_from_filehandle(batch_h):
+                dispatch_batch_statements(statement, counter)
+            batch_h.close()
+        except ValueError as e:
+            click.secho(str(e), err=True, fg='red')
+            sys.exit(1)
+        sys.exit(0)
+
+    if not sys.stdin.isatty():
+        batch_h = click.get_text_stream('stdin')
         try:
             for statement, counter in statements_from_filehandle(batch_h):
                 dispatch_batch_statements(statement, counter)